+++ HISTORY.rst
@@ -5,6 +5,7 @@
   links.
 - Use Select2 widget for the document type selection form.
 - Backport the vertical main menu update.
+- Backport workflow preview refactor. GitLab issue #532.
 
 3.2.5 (2019-07-05)
 ==================